disable imgui and move fps counter to UI

+++ src/ecs/components/io_components.hpp
@@ -9,7 +9,7 @@ struct WindowState {
 	int height;
     bool isRetinaDisplay;
 
-	float fps;
+	int fps;
 	int numFramesThisSecond;
 	time_t currUnixTime;
